Add olm.skipRange and minKubeVersion keys to csv
This was reported as a
[BugZilla](200~https://bugzilla.redhat.com/show_bug.cgi?id=1804889)
fill the missing olm.skipRange and minKubeVersion variables are in the nfd clusterserviceversion.yaml

miniKubeVersion set to 1.17.1
olm.skipRange set to 4.1 - 4.5

minKubeVersion is the minimal KubeVersion required by your Operator.

olm.skipRange is the scope from which your clusterserviceversion can be upgraded from.

Thanks to Anping Li for bringing this to our attention.

Signed-off-by: Carlos Eduardo Arango Gutierrez <carangog@redhat.com>