Updating documentation in 'main.go' for 'metricsAddr'
Prometheus is used to scrape data (metrics) in relation to NFD resource usage. It is
important for contributors to know and understand what 'metrics bind address' is and
how it relates to this data scraping. It is also important for contributors to see
how the metrics scraping bind address relates to the manifests under './manifests'
and './config'.