Adding documentation for NFD related structs (types)
Adding documentation for structs defined in the two files in this
commit. This documentation describes the fields within these
structs so that contributors know what each field means and what
each field does. These structs are used quite a bit in other .go
files, and so it is important for contributors to be able to
understand these structs.