Convert common code to generic function that works with any resource
The code logic for interpreting condition results is repeated across
multiple functions. The idea is to create a generic function that
works with any resource, as each resource technically has different
types for different attributes compared to other resources. For now,
the results are returned as rstatus.isAvailable, rstatus.isProgressing,
etc. instead of just returning 'rstatus'. This will be fixed in the
final code.