Change NodeFeatureDiscovery struct attributes and add checks for availability
The NodeFeatureDiscovery struct was using 'ClusterRoleBinding' and
'ClusterRole', when in fact it should be referencing 'Role' and 'ClusterRole'.
Moreover, add a function to update the availability of a given resource.

For now, the status checking functions are not in order and will be put in
order soon (i.e., starting with SCC, SA, etc. resources, not with 'Service' or
even 'DaemonSet'). This commit is purely for checking DaemonSets for now.