Split 'DaemonSet' attribute into Worker and Master. Also fix parsing.
Prior to this commit, the NFD instance (object) only had a field
called 'DaemonSet', but this is not correct because we have two
daemon sets -- a worker and a master.

Also add code to parse the daemonset JSON outputs.

This push is very much a WIP and it will not compile right now.