Report NFD operator status in the nodefeaturediscoveries/nfd-instance… (#183)
* Report NFD operator status in the nodefeaturediscoveries/nfd-instance CRD

Prior to this commit, the NFD operator's status was not reported
anywhere. With the addition of this commit, the NFD operator's
resource status is now displayed in the NFD CRD when viewing the
output of 'oc get nodefeaturediscoveries/nfd-instance -o yaml'.
If all NFD operator resources are healthy, then the status reported
in the CRD is listed as "Available." If one or more resources does
not exist, then the status of the operator is reported as being
"Degraded," and the reason for the degraded status is reported.
Finally, resources that have been deleted but are in the process
of being recreated are listed as "Progressing."

Merge branch 'master' of https://github.com/openshift/cluster-nfd-operator into nfd-status

* Remove unused imports

References to the "errors" and "reflect" packages were removed from
the code because they were no longer needed. However, the imports
still remainted, so those are now removed.