Add initial code to introduce finalizers to the operator
Add code that adds finalizers. Also update the
controllers/nodefeaturediscovery_resources.go file to include
"getter" functions and "delete" functions to streamline the
process of getting and deleting resources associated with the
finalizing process.

Note that the getters and deleters will be useful for the NFD
Operator's status checking logic, as right now, the NFD logic for
the status checking is replicated by the getter functions.