Fix logic for determining if an NFD Operator resource was updated
Prior to this commit, the controller manager did not properly
check for updates to the NFD Operator's resources. It was only
checking if the resource had an "old version" that was replaced
with a "new version." However, we must check if the old version
and new versions are indeed different. If they're not, then no
update should be performed.