Fix logic for disabling/enabling leader election
Prior to this commit, it was impossible to disable leader election
because `LeaderElectID` was set to a specific value. In other
words, setting `LeaderElectID` in `main.go` automatically enables
leader election, even if you manually modify the relevant manifests
to set the leader election to false. The only way to actually
disable leader election is to create an if-else conditional that
creates a manager *without* the leader election variables passed
into the constructor.