Fix predicate functions to correctly see if a resource has changed
The existing predicate function did not correctly assess whether
or not a resource had indeed changed. It only checked if the
resource had "changed" by seeing if the reconciler attemped an
update, but it did not check if the resource's Spec, Annotations,
etc. had indeed changed. Thus, we were getting a lot of false
positives when we tried to check if a resource had changed because
the reconciler, by default, attempts to update one or more
resources regardless if the resource(s) need(s) updating, so there
will always be an update event even if nothing was actually
updated.

It is also important to add CreateFunc predicates for each resource
because we want to prevent the reconciler from crashing if the
reconciler sees that an NFD instance has changed. (i.e., If we
delete an NFD instance and recreate it, the reconciler complains
that the resource has been changed and needs updating.)