Consolidate code for finalizers and CR status updates
Currently, the same logic is being used for updating the NFD
instance CR, as well as adding and removing finalizers. However,
the code for this logic is repeated. Thus, the aim of this PR is to
consolidate the code so that the logic is not repeated, but instead
reused for all situations.