CSV: fix kube-rbac-proxy image reference
This PR fixes the kube-rbac-proxy image reference used by the NFD
Operator. Without this fix, the image is referenced by an implicit
`latest` tag, and this breaks disconnected deployment, which requires
a SHA image reference.

`registry.redhat.io/openshift4/ose-kube-rbac-proxy` is the string
[used](https://github.com/openshift/cluster-nfd-operator/blob/c8d4be46e46622f1f29479cedac4a020d73bc08b/manifests/image-references)
in `manifests/image-reference` for the `kube-rbac-proxy` image. This
string will be replaced (by ART) by the image+SHA.