Do not set explicitly image pull policy to Always
In Telco far edge use case, a cluster might have limited management bandwidth.
Upgrading such a cluster is expected to be time-limited (capped by a maintenance window) and will involve container image pre-caching on the node.

Image pre-caching solution for nodes with limited management bandwidth relies on the assumption, that workloads can use locally stored images without contacting a registry. This is possible if the container image pull policy is set to "IfNotPresent".
The NFD operator is using Always as the pull policy, we should delete the policy and have the default behavior.

From https://kubernetes.io/docs/concepts/containers/images/#imagepullpolicy-defaulting

When you (or a controller) submit a new Pod to the API server, your cluster sets the image pull policy field when specific conditions are met:

if you omit the imagePullPolicy field, and the tag for the container image is :latest, imagePullPolicy is automatically set to Always

if you omit the imagePullPolicy field, and you don't specify the tag for the container image, imagePullPolicy is automatically set to Always;

if you omit the imagePullPolicy field, and you specify the tag for the container image that isn't :latest, the imagePullPolicy is automatically set to IfNotPresent.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>