Requeue explicitly after adding finalizer
According to the Kubernetes documentation, .metadata or .status updates
do not bump the generation of an object. The NFD controller's update
predicate checks fot the difference of the .metadata.generation field of
the NFD CR. Adding a finalizer (i.e. .metadata update) to the NFD CR,
updating its .status and returning while not explicitly requeuing the
event, leads to the NFD CR not being fully reconciled after its creation.

By explicitly requeuing the event after adding the finalizer, we make
sure that the NFD CR reconciliation will continue as expected.

Resources:
- https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definitions/#status-subresource

> The .metadata.generation value is incremented for all changes, except
> for changes to .metadata or .status.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>