Fix the absence of status from NFD CR
Since moving the NFD CR status to use the apimachinery conditions,
the reason fields has become non-empty mandatory field. Since it
is not set, the update to status is failing, and therefore status
is not shown , since it has never been updated.
This PR adds Reason field values to all the conditions

OCPBUGS-7023