Fixing prune job implementation
No need to explicitly delete the prune job, since it will be deleted by the K8S scheduler once the NFD CR is removed from etcd.
Additionally, if the job is explicitly deleted, then the code will attempt to recreate it a second time while the NFD CR is being deleted.