Getting rid of pkg/utils
pkg utils is used for solely for getting the WATCH_NAMESPACE env
variable . We don't need a separate public package just for
that functionality. Moving the function into the main. In addition,
since WATCH_NAMESPACEmust be set in the deployment, the operator will
fail in case the variable is not defined, instead of just logging a
warrning