Removing handling of TopologyUpdate scenario
TopologyUpdater was added due to NUMA team intention to use NFD for
topology resource reconciliation. They have not moved to using NFD, and
therefore we have 2 operators trying to deploy the same CRD, which
causes collision.
This PR removes deployment of NRT CRD and removes the code that creates
an TopologyUpdater daemonset.